requirements.txt

matplotlib
numpy
pandas

scipy
tslearn
scikit-learn
transforms3d

seaborn

umap-learn
